Move tests to a dedicated file `tests.rs`
